Fix broken links
Header (lines starting with #) create anchors with their name and an identifier. This identifier is used to link to them. Multiple headers with the same name append -n to ther identifier. 

To retain the copy & paste feature of this readme, the desription headers in the chapter "Under the Hood" have been appended with a "- " to generate unique names.